+++ src/controllers/movie.controller.js
@@ -1,0 +1,3 @@
+import movieModel from "../models/movie.model";
+
+export const read = () => movieModel.find();
